requirements.txt
certifi==2018.8.24
chardet==3.0.4
defusedxml==0.5.0
Django==2.0.3
django-extensions==2.1.2
django-filter==2.0.0
django-ordered-model==2.1.0
django-social-auth==0.7.28
django-tuauth==0.1.2
djangorestframework==3.8.2
httplib2==0.11.3
idna==2.7
Markdown==2.6.11
numpy==1.15.1
oauth2==1.9.0.post1
oauthlib==2.1.0
pandas==0.23.4
Pillow==5.0.0
psycopg2==2.7.5
PyJWT==1.6.4
python-dateutil==2.7.3
python-openid==2.2.5
python3-openid==3.1.0
pytz==2018.3
requests==2.19.1
requests-oauthlib==1.0.0
six==1.11.0
social-auth-app-django==2.1.0
social-auth-core==1.7.0
urllib3==1.23
xlwt==1.3.0
